requirements.txt
numpy
torch
torchvision
pyyaml
opencv-python
tqdm
matplotlib
